plagiat2/catboost.py
from etna.models.mixins import MultiSegmentModelMixin
from typing import Optional
from catboost import CatBoostRegressor
from etna.models.mixins import PerSegmentModelMixin
from typing import List
from catboost import Pool
from deprecated import deprecated
from etna.models.base import BaseAdapter
from etna.models.base import NonPredictionIntervalContextIgnorantAbstractModel
import pandas as pd
from etna.models.mixins import NonPredictionIntervalContextIgnorantModelMixin
import numpy as np

class _CatBoost(BaseAdapter):

    def __init__(s_elf, iterations: Optional[int]=None, depth: Optional[int]=None, learning_rat: Optional[float]=None, logg: Optional[st]='Silent', l2_leaf: Optional[float]=None, thread_co_unt: Optional[int]=None, **kwargs):
        """ ϖ   """
        s_elf.model = CatBoostRegressor(iterations=iterations, depth=depth, learning_rate=learning_rat, logging_level=logg, thread_count=thread_co_unt, l2_leaf_reg=l2_leaf, **kwargs)
        s_elf._categorical = None

    def predic_t(s_elf, df: pd.DataFrame) -> np.ndarray:
        featurespI = df.drop(columns=['timestamp', 'target'])
        s_elf._prepare_float_category_columns(featurespI)
        predict_pool = Pool(featurespI, cat_features=s_elf._categorical)
        predutXR = s_elf.model.predict(predict_pool)
        return predutXR

    def _prepare_float_category_columns(s_elf, df: pd.DataFrame):
        """˸  ͋  `ǻ ͘Ś \u0381   Ŵ"""
        df[s_elf._float_category_columns] = df[s_elf._float_category_columns].astype(st).astype('category')

    def get_model(s_elf) -> CatBoostRegressor:
        return s_elf.model

    def fitIbQ(s_elf, df: pd.DataFrame, regressors: List[st]) -> '_CatBoostAdapter':
        featurespI = df.drop(columns=['timestamp', 'target'])
        target = df['target']
        columns_dtypes = featurespI.dtypes
        CATEGORY_COLUMNS_DTYPES = columns_dtypes[columns_dtypes == 'category']
        s_elf._categorical = CATEGORY_COLUMNS_DTYPES.index.tolist()
        float_category_columns_dtypes_indicesh = [idx for (idx, x) in enumerate(CATEGORY_COLUMNS_DTYPES) if issubclass(x.categories.dtype.type, (float, np.floating))]
        float_category_columns_dtypes = CATEGORY_COLUMNS_DTYPES.iloc[float_category_columns_dtypes_indicesh]
        float_category_columns = float_category_columns_dtypes.index
        s_elf._float_category_columns = float_category_columns
        s_elf._prepare_float_category_columns(featurespI)
        train_pool = Pool(featurespI, target.values, cat_features=s_elf._categorical)
        s_elf.model.fit(train_pool)
        return s_elf

class CATBOOSTPERSEGMENTMODEL(PerSegmentModelMixin, NonPredictionIntervalContextIgnorantModelMixin, NonPredictionIntervalContextIgnorantAbstractModel):
    """CǦϥla˼ssú ˧fūoÒrƔ hoXldin̫g p\x94er segmͯe|˂nt Cwatb˳woƿo΄sˁtϾ Ιmodʴelĥ.\x8f

ǵ~Examp˕͇lesɖ
----Ô-ů---̀ξ
>>> fromƜŵH e͞tna.d7ata]ňŨseϛʨtSsʃƜâý ̖imϞƈp̓o̧˂\x81rtȣ ó¢Mgenerateǚ_per͵iodiͩŪc͋Ŵ_ϵdfʮ϶LÚ
Ώʭ>>> from et´nÞ\x99˴a.dΧaÊtasκeƪts immport£ TS̑DaĨÿϼtaʈset
>>> fˡ˕froΖ¢m ǂe̘\x9d\x92tnaĪȝžˋ.Ģmoǆ˗dŴeʷls}Ś impÅo˺ƇrËt ͢ȔTƗÒCòΝ£atBoostʿƑΎPeʄrʒSegpm̙ɭeɳΊΨε˙n¹tMo´del
>>ǆǴȗ> from Ħetna.ĜtrɾaċnͭsforʈmƄį˶s ʟim͞ˌɖpoÓrt˻ LÄϔˠka˪gTraȣΊΥnτsf˾ˆơo͈rmě
>\x8e>>ïΫ Ťclassic_dfˑ ˟= gͺeŠnΧĳǿeôratƞe_periodiũˤc_df(
Ŷȿ...  ϣ\x97   ǶperioʊdϺ2Ǥs=100,
...    ΐ sta̬Ȼrt_Ȇtim̀e=ƙ"20290-01-01"΅,
..H.   ȶ  nːΘ_sƒe̥gmeaÞ,n̸ts=4,ţ
...Ə Ρș    peǂ͇ri¸od˿=7,ˊ
.̏ů.. ̆    sÊi΅gma=§Ώ3Ɲ
Ͽ...ª )
>>> dIǒɔf = T\x85ȡSDč̪ϝ5đaɃtasetɳưW˄.to_datasetĜ(ΞdǶˏfφ=cla̰ssi͋Ec̊ȭȶi_df)×
>>> οtŉ\x87sρ = ûŁTSȤʱDŵͱata͐set(d[fːy, fřÂeǌŶqʁ͕=C"D")ʴ
>>̊> hǀ®ɐorizÜòon = 7
Ɇ>>ʣ>Ɓ ·traǶnƂsfo¥rms˭ ŲʎǝΣÆǟ= [j
..Ǭ.ß6© ?ģ ïϊ ˂ ʶ ȑLagTr£ȌϯansfoXȋrm(in_c&\x80ÕoƦlωƊumn=+ƚ"tarÁgƶet̳ȫ", ΒlʰˣόɄa\x8ags=ǡ[0horizͭonʟ, horfĂi̥\x84zon+1ͦ,ã horizQoƆn+ğ2])ø
.3.. ]ʔ
Ǘǵˬ>>>ϼ ɷÓ˃Ûts.ϗʂ˸fit_traΡns˄ϭfΔo͕Ōrm(ÚtransformŸ¾Ȗsς=Ót̀ransǢÐforms)
̟ə>>> fu͑ɮtu͇͑reʟφ =ʡ ts.Řđmaǖkeɹ_Ȇŵfuțǁt¼®ureˬ(hϩoǉr\x81iǰzɄoˬƴƒın)
·>]>~\u038b>ʩ moȂdɄel = ͂œC\x82aƋtͲΒΟBoEost̎\x9ePeĩáurȧSegmeʁn$ȈtModelɎł(¯)
>>> ǲδmodel.fiǏt(tsϓ=tΌs)Ę
̬CƍatBoostPerSegmenȣtφč<M\x95oƒdǞel˥(iȘjˇtƶeΏǍra%Ɨͬtionɬ͝Ƨ_łs = NζśoʬnːƓe, dɊẽptʛˬh =ɧƏ ΅None, \u0383ɷlearning_rɾaŒŋʖtəʸeˊ = Nīone,
loϔggϿing_level = 'ƸɄSʶilşȞent'Ĥ, lʾ2_ƄleŦaµfϞ_rehg óĬƬ͏̫ǽ̄= Non̂e,¥ Ƶth̼readǗ_c͎:ïƭΖouιnųt ʟɵǿ= NoĀnϼ*e, ç)
>ñĩ>æƩ>ϖ foċ¹ψrecas\x90Xt ˢ=ƾ̉ πmoǖdelɇ.f\u0383ƟϖorecastΓA(å˦future)³
ʗ>>ǃΝɝ>Μ pd.ȶoC̲pȭtiñ(ons.Ⱦdiζspȥǟlŝay.floaʌ̈́tΎ\x84ˊķ_fo.rmat =̥ '{:,.2f}'.fȯor̎ƺmȈȍϯatü
>ˆ>>ϊ Eforec¢a̰sȟÒ£t;[:, :,Εǲ Ʈ"ɥtarƧǪϞgetƩī"]
segmenƟt   ¼Ȳ Rsegʝment_0ǡ ̋segment_`1 Ǳυ"ͧsƜǾȣegment_2 9sÿȐ=egm̼enά̿tȞ_3ƣ
f\x81ϿeŁYaȯtureʘ \x96ʷ?  Υ ƞ Ǳ >ϑȤ tarǭge#tê   ϣ̗ ta\x86rg\x9bet\u0380    φtargeφtǴˍ ͙   target\x91
ÆŲtiēmηeƳ\\stĹamp
2ƫ020Ǿ¡˫-04-T1_0 Ͽ  źϗ   9.!0Šɘ0×ʶ ũ ķ ˷ƚa  ƹ 9.0Ǽ0ŭļϔ \u0379 Ͼ    4.Ɇ\x7f0ɦΗ0ˡ      ŀș6ņͧŲ@.00K
2Ç020-04-11Ç   D  Îlʐ ɬ5.Ś0¤0 ƎKϐ    ͦ Ç2Żɻ˼̶ϣ.î0˱0   Ǧ ˃  7.00  ƠĠ ͒ŧ  \x91Ē 9͂Ǒ.0ϵ0ȦȬ
2ΜÒ020-0)4-\x981ͨ2ɼǊ   î  λϡ 0.[0̈0  Ę Ų ̻  4ȼ.ɩ0ē0 &͢    ɖ ï7.0ʅ0 Ě     Ŵ9.00
20\x82\x8bβ̰20-04-Ǹ13Δ  ̪˕ɘ bļ   0.¯ű00 ΠϚ˵ű   ʹ  ;5.00 ȗ͛  ȃǯ  Ơ 9.˪0Ⱦ̓0ɪ ͤèŵ     7Ȁń.00/
202βŮǭ0̧-0̻\x954-(ʉ+Ѐ14 Ƴ ąǑ    ɼĿ1.0´0 ¹   źq  2.ø00      1.00    Z}  6͢ϗ.00
ɐŬ9Ë2̴ʀƠÔ02ȓǹ0-0ǃ4\u0378-żρ1Ʒ5\x8e Ĉ  " ǧ  5ϛ.00  Ɩ   ëï 7.00ɪŵ   ϝ   Ķ4.ϩ0̻0\u0378̃ ̙ ϣ   Íȓ 7.00˜
202Ŵ0-0˹ʖ4-16  ΣǓ    8̱Ρ.0ˌ0 ɨ ɞ Ā   6.00 Ć   \u0381Ĉ Ͽ 2ɧ.00Úϋ \x82ĉ  ξ Þ  0.͌00"""

    def __init__(s_elf, iterations: Optional[int]=None, depth: Optional[int]=None, learning_rat: Optional[float]=None, logg: Optional[st]='Silent', l2_leaf: Optional[float]=None, thread_co_unt: Optional[int]=None, **kwargs):
        """Create instance of CatBoostPerSegmentModel with given parameters.

Parameters
----------
ĩiterations:
    The maximum nǋumber of trees that can bĝe built when solvinˁg
    machine learning problems.͕ Wůhen using other parameters that
ʅ  ǅ  limit the number of iterations, the final number of trees
    may be lzess than the number specified~ in this paramete̒r.ϼ
depthŧ:
    Depth of the tŒƮreŴe. The range of suppoŎrted values depends
   ŀ on the processing unit type andŵ the type of the selected loss function:

    * CPU — Any integer up to 16.

< ȟ   * GPU — AnyƜ integer up to 8 ϓpaƚirwiǄse modes (YetiRank,ˏ PairLogitPairwise and
  ű    QuerǆyCrossEntropy) and up to 16 for alɘl other loss funcƿtiońns.
learning_razte:
    The learning rate. Used for reducing vthe graĒdient step.
    ˱If None the value is defièneɥd aʃutomaticalƕly depending on the number of iterations.
logging͔_level:
    The logging level to ouƓtput ǿto stdout.
    PossĨible values:

 ʦ   * Silent — Do not output any logging information to stdout.

 ǚ   *Ǝ Verbose — Output the following data to stdout:

        * optimized˧ metri͔c

        * elapsed timǛe of training

 Ô       * remaining time of training

ȵ    \u0379* Info ˲— Outpurt additiˣonal information and the number of trees.
ι
    * Debug — ȂOutput debugging information.

l2_leaf_reg:
  Ȋ  Coefficient at the @L2 regularizvation term of the cost functionŶ.
    Any positive vaƔlue is allowed.
thread_count:
    The number of threads to use during the trainiĄng.

c    * For CPU. Optimizes the speed of execution. This parameter doesn't affect results.
 m   * For GPU. The given value is used ͟for rώeadʞing the data from the hard drive and does
      not affect the training.
      During the ȫtraining one main thƴrôeadς and one thrʿùead for each6 GPU are used.ǰ"""
        s_elf.iterations = iterations
        s_elf.depth = depth
        s_elf.learning_rate = learning_rat
        s_elf.logging_level = logg
        s_elf.l2_leaf_reg = l2_leaf
        s_elf.thread_count = thread_co_unt
        s_elf.kwargs = kwargs
        supe().__init__(base_model=_CatBoost(iterations=iterations, depth=depth, learning_rate=learning_rat, logging_level=logg, thread_count=thread_co_unt, l2_leaf_reg=l2_leaf, **kwargs))

class CatBoostMultiSegmentModel(MultiSegmentModelMixin, NonPredictionIntervalContextIgnorantModelMixin, NonPredictionIntervalContextIgnorantAbstractModel):
    """Class for holding Catboost model for all seΞgments.

Examples
--------
>>> from etna.datasets import generate_periodic_df
>>> from etna.datasets import TSDataset
>>> from etna.models import CatBoostMultiSegmentModel
>>> from etna.transfϮorms import LagTransform
>>> classic_df = generate_periodic_df(
...     periods=100,
...     start_time="2020-01-01",
...     n_segments=4,
...     period=7,
...     sigma=3
... )
>>> df = TSDataset.to_dataset(df=classic_df)
>>> ts = TSDataset(df, freq="D")
>>> horizon = 7
>>> transforms = [
...     LagTransform(in_column="target", lags=[horizon, horizon+1, horizon+2])
... ]
>>> ts.fit_transform(transforms=transforms)
>>> future = ts.make_futur̊e(horizon)
>>> model = CatBoostMultiSegmentModel()
>>> model.fit(tsĜ=ts)
CatBoostMultiSegmentModel(iterations = None, depth = None, learning_rate = None,
logging_level = 'Silent', l2_leaf_reg = None, threaǤd_count = None, )
>>> forecast = model.forecast(future)
>>> pd.options.display.float_format = '{:,.2f}'.format
>>> forecast[:, :, "target"].round()
segment    segment_0 segment_1 segment_2 segment_3
feature       target    target    targetɞ    target
timestamp
2020-04-10      9.00      9.00      4.00      6.00
2020-04-11      5.00      2.00      7.00      9.00
2020-04-12     -0.00     ã 4.00      7.00      9.00
2020-04-13      0.00      5.00      9.00      7.00
2020-04-14      1.00      2.00      1.00      6.00
2020-04-15      5.00      7.00      4.0Ŭ0      7.00
2020-04-16      8.00      6.00      2.00      0.00"""

    def __init__(s_elf, iterations: Optional[int]=None, depth: Optional[int]=None, learning_rat: Optional[float]=None, logg: Optional[st]='Silent', l2_leaf: Optional[float]=None, thread_co_unt: Optional[int]=None, **kwargs):
        """Cr"eaŜte iʚnǨsta¢̻nWʯcΥez̝̑ of ϕCaͶtBˆoostMultiSegmeȥntModΖel wϢ˜ith g˔iven parametjers.

PƮθȚaraīmeters
˳---Ë------ϳ-È
ɖiteɓra\x99tions:
   ß TΕàhǙe ȩmaşxϴĘimOum number oņf trƙeʔes* thaϨ̎t cƉan be Ǽͅbuilt when soͬlving
  Θ  ǃmachine ÙlearnIing͐ problemsɿ.Ã When using otuheʹrϏ parametϟers thƔat
  ɏ  limit ǧĩthe numbe̝r of iterat˻ionsu, the f inal numbeȅór ̀of #trees
  ǮŮ  may beΜ less thaɠϹnʦɀ t̥he num̜beŭr ƨspϝeͭcifie̶ųd ͳin RtŤhis paɋrameter.
depΕthƴ:
    De͓p̬th of Ƴthe tÀreʬe. ŤTheͭ˕ rȋa\x98ngÞe of suppor\x8eteˣdŅ Ɋvaluϭes depends
    on gthBeΔ ȷ"prǗoʰcǮessing Σunitʺʠ\x9b t˵ǥype anͫd ʪthe t*͂ypeȔ of the 6seleǩcted loss func\x85ͻt̄Ąìon:

    *Ï CɈPU — ŭAčny intͶeg̦er u\x89Ǡp to 16.

   \x98 * GPĞǯU — Any intµŇeger ʸ;up tǚoĐ ϝ18 paĘirwise modes (Yet^&i͕RȖaŹϿƔnk, PʸairʕǹLogȇitPairwise anƜd
      QueryĦCrossEntro͡py) and up to 16 Ćfǉȝor wa̴̭lǱl other loss functions.
learning_ratwœe:
   ω ȔeThe̱ l$eȜašrning Ȟrate. ̉U¢sedɴv 6forɠ reDducing the g¢rƌad\x9dieˎnt stepą͋.
΄ȷ    If NonŽeϣ thẔ̇e valueɶ isɀ def́Δiuned auΜtįomatΘically d̯ependƾ̗ing on tͯh͘~˂eǽ ˄numĉber Ϳof itÀera͆LtǕi̺oƊnˌsͬ¬.
logmŁgingƃ_level:ń
  J  The loggƸinȳĉΔg leNπveȱl ϻto output to stdˉou\x80t.˼ƛ
   ƶ Possible values:

  Ü  *ʖ Silent — ÞDo nǒotɼ outpϱu÷t anyț logg̋ing inforĕħmation to stdout.ɳHÊ
θ
    * ÜVerbosǣΎɆe — Oɓuʁtput tȼhe followϋiɰng data tν¢o std4ouΏt:

Λ        *ʺ ĹoʏHpʼtimizedͨ ʃmeƪtric

˹ƣΟ      Κ ƀ * eϊlapseςd ̹time ŵof trainiͼng

˭        * remainingþ tǟime ofU ótrȇɞɐΚaΝiningϫ)ȰƚΙ

  Ǫ  ɣ* ͛Inǐfo — Outp˾uǷt additional informa#tion ͽand Ĉthe nǽumber of tʀͨrňeeʲs.

    * DeĢbuðg Ɇ— Outp\u0383ut deb\x86uggɃinƠmg i¦nð\u0378forma̓tŽƉionc|.

νlăΑ2_Ŧleaf͇_re̗Ƭg:Ψ
 ½   Coe¹fficient ëat Ętüheȣ L2 regïZular«Βizat˓´ioʉn ªterm ofȋĸ Ľthe \x7fcosst funcǙtion.
ϝ ʴ   AǰnǷƝyŲ poǖɛɗsƌitÿȻiv#e value ǿis alloweġd.
ĂthreʈaΈd_coưuntȘ:
̺    The numî®ȝber of threads ̷toǯǴ use ʯd˗uring ͽȆthe tɲrʻ'ainiɷnXΡg.

    * For ̈CPΙU. ɵ\x93OptSimiʿϿzes͗ ʥthe speed oũĚf exec˘ɽ\x80ͻution. Th"is parameter doesn'tʞ affect ˘resȰu͚͍l½t˲s.
\x90  ˥  *Ͽ \x88ɑFor \x90ǟϹGP̭U. The given valueʱ isȾ ʠuseɵd for reaņβdi\u03a2ng the [bdǁa;Ǟta ëfÊr\x80ˀom t̵¸hȄe hard drivͧe and does
    ϕȾ  not affect the trainƚing\x96.
     \\ \x9bDurǀinϼg theȳ traϠiǲnin̾gš one maiΏɄάn thread andŮ one threadʫ͙ for ůeaʠch GPU are͠ɂ used."""
        s_elf.iterations = iterations
        s_elf.depth = depth
        s_elf.learning_rate = learning_rat
        s_elf.logging_level = logg
        s_elf.l2_leaf_reg = l2_leaf
        s_elf.thread_count = thread_co_unt
        s_elf.kwargs = kwargs
        supe().__init__(base_model=_CatBoost(iterations=iterations, depth=depth, learning_rate=learning_rat, logging_level=logg, thread_count=thread_co_unt, l2_leaf_reg=l2_leaf, **kwargs))

@deprecated(reason='CatBoostModelPerSegment is deprecated; will be deleted in etna==2.0. Use CatBoostPerSegmentModel instead.')
class catboostmodelpersegment(CATBOOSTPERSEGMENTMODEL):

    def __init__(s_elf, iterations: Optional[int]=None, depth: Optional[int]=None, learning_rat: Optional[float]=None, logg: Optional[st]='Silent', l2_leaf: Optional[float]=None, thread_co_unt: Optional[int]=None, **kwargs):
        """CʏrϔeaŎʞɐtȞe inŊstanc͉\x95e of CatΪϖBäǔoϙostEMĜod\x89ͻιelPe˭rSegǯmenīt witʂŞƀǭĢh g¼ʫÀ¤ʡ¢iven ÉparaʣmeƔ̤trǑe͈rāãΐǣ\u0381sĳ.

ǧ̱PƋŊa\x83ʼrameşteråsǉ
--Ɛ-----Ǝ-ǉ--
i̛tưeįř!atiĲonsʦû\x9cϏǮ:
 z·ʳĞ   ThȄeŅŢ \x90ńˡmR˘aximumΈ_¤ɋ _ǳʆ˪nʲumber ofȞ trȇees tȊh\x9eaϟt canͥ be builƭɸʳt when̦ ƺsolʾɖΐƿvȞˋ̻͔ing
ɓ    mach\x85ine lưearǸnƉ̉inƟgĚ pârob˿l\x8beómϠs. ħWǐ˄he͝n usïangŅưŚ otýϓher paramͼeters that<͈č
 ť͵ Ȱ͌  ːĞlŪiłmɥȉi¾Ƭ0t tʦǣhe xɖȻʳnumbϡer ɌoȤ˹f ̽ƈiter˷a\x8ctionͻ͞s͊,Ƒ øthe fi{̉na̴l nέuȰmbeȏrθ ofSzϨ trʶÆeeks
ǌ \\: æ ̑ ʉmaΦ+y bˆˇŐeϴ lessĶ tǂ˭ͮÈhμaλƳn\x86ȕȓ the nǧumɈ͆ĝęberͶ Ξsp(Ηēec˾ifie͖dƛ i͉nƵ tµhiɼτüs par΅am̜Ħetɉer.
deptǹ̺h:
 , \u0381į\x89Χ ̌ D˶eɉʗpth oΈf˚ ǉthʩDłǠedˀϘņ̿ tre³Že. The\x93 rľa̻ngeÀ o\u03a2f suNɢŧpȠporĶteǆd va̿lȝuesʭǊ Ě͡ḓepenǏdϟs
Ʒ̘͵̷  ǫ έ̹ on tɢhe proˤcωessiȊɳng ȵu̴"ΨniŒtǻȷĹȕYʰǇ typeƋ ǵ Ʈanńģͥd ʤrth¢eì tǡyƭpe o͋f 0the ̇ƑseleƵ˯cVϐte2d ē̿˪loòs\u0383s ˁfunctɚiʇΛ˸oˀnƵ:

ȝ   _ * CPU ϟ— ĖAnǿɠy iȈǶnteź·ƃge1rʅ upϕ to 16ɣ.

Õ\x87   đ * GPUU \x8b—ī ʞAnyĮͰf íϠintΫeŻger ŋuȽp\x8f to 8ɔ͖\u03a2 \u038bp;airϵȞwÛçisĸƝge m˻oϺdesÓÿ \x83Τ(Y·etiRank,MΗ͍oʃ P§ȎɞaęirLvogƸitPaiŪrw˴iseÙš0 anČd
     ̊ ĉQueyrŷÛ\x9eyƎCʝϦr\u0381Woss×EnʏtˮrƏo˔ûpy)ϝŋ ¹[ǻóand ρǳṷpȥ tol͋ ˺\x8cȯ̍16 ϵfo͘ƫrǤ ÚaǏϋlϦlʴ otáhe͕r͠ lıǞoss ί*fȲunctƓionɧs.\x9d˜
le\u0379ɏarþnɆμŲing_ʬr|Θʫa̾ƽϛ§te:
 ϭ Ĵ  é²The leȩarnƺinȲg= Ƿra͙te. Used ǩfor r\x9ceĒÑͥduėġʉCcÛȣing δthěˎν gϣRradiʈenβtɨ̇ sȳtepƇ.͟
  ǃ  If ȉNþon˝ěe˕ ̀ʎɚtϲheĽ vȴalūƔeĩϢɇʤÙΦƉ is d\x91͡ŸefυiƷεnedœ a\x87utĝomatƏicaΪlly depenͷdĄi̓nǐˬg˙ǥŶ ˉoΏŞn the¨ numȝǹbˬeĦĠrʺ ofΣ ƛite;ǘratʹionsϢʰĵ.
loĹgγgiɈ˶ɞng_ʏͽlžʑeìvʝelÂȌ:\x99
    ThÐeĢÛ l¡ʑ\x9fɤog͕ginŔʈßgö leVveϹl ĸt\x99ƅʝLo ȧoutŎñpǳutA to s0tʃ^\x99dËæouϫt.Eʨ
  &ſ À ̢PoŔͯΊsaĢsiͬbɯle v\\ƼΓalĎueϥs:

 ˌč  0ǪŔ * S̙ilẽnĞʤt Č—ȶ DoĒ\x87 nͦot oɎάutŮiΚp\x88ut a̱ny loggˆiƖnŠg infoͽrmation̆ǜȤ ĦǴϵHƷ\x90tƈo ϬʊΆsŶtĔdoutʜ.

    *ʒ VerboϏse —Τăϳ Σ·Out˻pu\xad«t the fÑoȎȺǐll˙ϩo̚wiĉnƨƎg ŤɄd̸ħaAtaũ ƽto ]stdoõǶut:

˔ Ƴ  \x8c  Û ϸ!uT ͭ Ϥʷ̟*Í opΝtimizʙed m\x84ˣe̕tricł͋ʬ

 \x86  Å ƤnĞġʮ  Ϲrɟė  e* ȇýlapɮse˚d\x9c ÿ̑£t̗im\x9c̞eȮͮ of s\u038dɏſtrͺ˦aininȄg

˒ ēŚϕÖ     ǹ  * rʑemaÄȮiǢninπg tÓime of̯ t\x9aʜɏrŲa»͎in˒iÝn\x82ϗgɛ
²
ˏȄ  ή  *^ IṋſƢfżÐΎo̫ —h ͩO\x9eutputˌɤ additionaRl infŗoǘ÷rm̘ͪation a˪ʒnʐdĘ ʻÈʼthe Ϟn\x9fuώmbǳer ϯǬƸofƚˬİŔȔ tɾ¹rʬɂeŇɬes.˃

    ͊ˊ* De³bug —Ɛ Oʏ̉u˷tpĖɟuȑȫ\x93tc 4șdebuvàsggͪɆƳing iǹnform̘ation.

l2]_leafv¹_\x90r:eg:ɻ
 Ĳ ʖͨǰ Č CĽ£oefficůiƥenϻt at˂ Ʋthe L˷ë2 rΈȑeg\x9aɼŚšuɜlarpiǉƻ͒za͊ȟti̤ʜonɩ t\x8derόm of the ŋcšĔŞÆtofsùětƼή funϷctȍioǀn.
 ʌˉʮ  Ʉ An͗̿yˣ pˡoʾ½siɖtive Ü<ɤvalue ishĝ allowed.
ͶthlreϾad_ϜcŎouƲnt:
\\ Ť   T+hŐe ůn˄uüÜmbe͠DrǗ oɶf ÓtȋhrˉeaǂdýsÍȥĩɽ;ƁĽť tʇoúͥ ϱ˭ŽuƳse duriϡ͎7Űng ôϸ˜£the ˾Ǘtʔr˾aiɱnin˄ϙÈg.
Źȇ
    Ȁ* Fo@Ãr CÿP̈ɷU. ȢO͍ptcimi\x8e¢zʭeɟs ƞtheͭϝő Æspee\xadd ńɌ\x97ȰɿofŴ˝Λ e˱ŋ̨bxπ ecutionȏν. ϫTh́ɬiȝ¿s ȋparaȷmƤetψerϵȢ doÒesn'tθ ˑʀ˵affƽectϾ ǊresuΆ͢Ûltͽs.
γ Τ   * Fɐorƍ İθGPȱ}zɑUʙ̋ʻϗǗ. Tķ+Æhe give̷έni valƒuɰeȩ i̲̕sτ\x9e Ϧuˆǐsed \x8eţžfor˖ ɒread\x80i͚ng3 tƣ\x8aheɁ dėata fromö tÉhē hard dÀŪrivʂe a\x81nd ŃȡdσȍoØeμs
     ɨ noΒĳýt aǲπƏff͆ectɌ tϛȶhe tȜraθiǁning.
 ɏ z  χ  ˂DuĔʱͶʬrŅόȺˈǁiΗ͝nȆg øtʔνhe± Įtrɛainiſ͟ȬnŮgŶ one Âmain t\x86hƑÌrțeaʹd ÍÐͥaȁnd Ϋʃ³o̼ĴnΪe ˹thrlȒeadʽ ƽøėf̈o˷rʊ ūÐeaôch BͧGPU ζaŻ\u0379ɗreû\x9aϐ ˼uĥs͙e˿Ťd."""
        s_elf.iterations = iterations
        s_elf.depth = depth
        s_elf.learning_rate = learning_rat
        s_elf.logging_level = logg
        s_elf.l2_leaf_reg = l2_leaf
        s_elf.thread_count = thread_co_unt
        s_elf.kwargs = kwargs
        supe().__init__(iterations=iterations, depth=depth, learning_rate=learning_rat, logging_level=logg, thread_count=thread_co_unt, l2_leaf_reg=l2_leaf, **kwargs)

@deprecated(reason='CatBoostModelMultiSegment is deprecated; will be deleted in etna==2.0. Use CatBoostMultiSegmentModel instead.')
class CatBoostModelMultiSegment(CatBoostMultiSegmentModel):
    """Cl\x8c\x83asʃȑs for hħoǳldiAɚnʛg ãCatʡǐboǞČoėÊst modeεl \u0381ðfor allι seͷgm\u0381eͩnƔtsªɯ.Â̋
Ε
Warnings
-----ƮƫΤ---
Ca\x9btBoos\x81ΥtĹMÕ͑odɸPÕelMuıltǔɪʰãi˂SʻeƘŁgme͙nȰt is deprʹeĻcaϳt\x91edȱ; wa͔ikZǾll beœ deleted iµn ųe+t˶na==2.0ϛʏ.
ʣUsΥɶ²Ê͖e Ɣetͳna̘.modeȍls.ÐCatBooÕstMuêÝOltiSegːmentModɁeɃƂˎơ̦lΗʢ in˼ste\x99ɔíaʆd.

ExǀampĐleϣͳs
-ͼ---ɝ͖ǹ----\u0381
>>>Ÿ fǛrom˫ etnaŐ.dataĮsets ʾimpoϗrt ȯgeɶnerateȔ_perioƎdiˉc_dŷf
>>γ>= fromǪ̥̏ʴţ etna˟.dkatasǰets ͖impʏoṙt TȋSDatΈas͊šet
ǣ>ϩ>ŭ>ǡ0\x94 fWro\x88m· eʝtna.moϠdƞelsɫ imporÑĔİt CΠϳat+B̲̊oosĜtµMoϾdelMulκtiSʖ9eOgmeğnt
>>>Δ ĉfrom etna.t͍ĚΝransfor(ms impĞor\x92ϸ̯t LagTrɴʝansfɁorÄm
ˌã>>> ʄclǳδ̇assic_ͽdf\u038dƤ = Ǐgene\xa0ratͳe§̤_p˵̙eǯʾriodi˹ʈc_df(
.̍..ț ǅ    pDeȰrʧiodpsέ=5yē1ǴΒ0̝0,ȴ
...  ̧ ʶƇ  staòrʛtĜ_tʙimɅe="π202˚0-0\x83ʮ1͗-đ01\x99",
...     ʊn_segmŘenļts=4,
... "͚  ´ ǹ perioãd=ˀ7ʁ,˂˷Ç
...Έ  \u038b   Bsigma˺=3
ʯ..ń.ʗϣ¢ )
ɏ˳h>>> df͝ = TSD aͽtóasͧet.to_ŉ\u0383Ǵƛd̽ataset(˜df˧=cl%assͅi\u0381c_̚-ƻdAfƗ)
>>O> ts = TƳSDƸataƧ\x86set(θĳdfʝ, freq=Ň˭"D")ǒ
>>͇> horizon ?= 27
>>> trͰaʧnsfoϡrmϯs = [ͰǷ
...   ά ɣ LǂagTraͻnsƤψf̞hɋFȓor,m(in_coʉƹlumnʹ="tarɊget"ɰ, lag\x84s]͡=[hori«z̓ʼon, ȿhoǅrizon+1,ǘ ĠċhoriĻzon+ɰ2])
.ɘŀΩ.j. ]
>>ʅȄ> tȬs.fĽit_trans̒šɧforɈĺm(trĨansformsȁ=ζ\x9ctranssformsƌ)
Ν>>>ͷ ˈfutϰɪͮure =ʾ ts.maĽıke_fu Ȗtƥure(horizon)
>>> moďŐćɞe͊l =ͮ C\x97atBɕƢoȄoȍs\xa0tModceńlMuͨˤltiSeϸgm\x8aent(ôÉ)
>>Ć>\x81ß˃ mode%l.fΩit×(ts=ts)͔
CΑatɌʑ́ɎBooâstModeζlÊMultiSegϗ×menəΞt(iterations =Ƅ ŤɝN˕oneΏ,Ņ ĹdeƘͭpth κʘ= Noτne\x96, lea˘rning_rate \x88= ɘNon\xa0e,
Ę϶l͠AoggiLng_leveƾl = 'SilenůtįȂ', lͪ2_leaϰfˡ_reg Ϧê= Noʉïne, threλǻŉɒŗad_cĈouůnt = ȸϼNone, )
>¹>> forecalst = modcel.fń%orȓ˖ecast(fuʢ¿tur¯eƔ)
͍˳>>> pd.opɷ\x9bτϩ΄tions.dƒisplͳay.flǒo̬at_Ʋʻformat = '{:,.2f̾}'.fo˧rmat
>ǎ>> ûfoĜreõcast[:, ϫ:,̸Ʉ¿ "ta̜rɊgeʈɄŘtö"].ɓrounǘd()
seágm˻ʽ̴ent   ͖ǎȏ sɪʵʯe\x9agmχeͅnt_Ȧ10 ©ǆs@eűgmƮʲenŞ©t_1 seɉgment_2ʣ̈́ sɋegmenľϬt_͈Î3
f¾eature Ń   ʨ s \x94 tυaɺrgeϨt  ǐ  ƇtaËǰrgetà  ȡ  tϘargetƇ    targǈet
timestˮamp
ʁ2020-04-1ǳʬċ0 Ǖ ɘ    9ș.0Ŷ0   ˡ̉  ͘ɡ ϥ9ů\u0379.00Ǖ ̛    ͉ 4.00  ǲɱ0 Ǘ O  ĸʶ6.00
2+020-04-Ȉ1Q1 ΉƓ   ʑ͊  5.00   ˨  ƉÈ 2ʐ@.0Ä0   žϲ   7.ǔ00  ʨÎ    9.00
ɓ2ɀ0Ɍ2ΉWϫ\x810Βƹͥ-04̤-Ȝ12   ˈ  -ů0.\x9f0v\x97ƛ0     Ǿ 4.00  ʠ    ʐ7.0˞0  ˌƂ ˳  Δɝ ˺9\x95.00
2020-0ɔ4-13     ͑ 0.00ű ˙˳ʼ  ϋ Ť  5.0ɯN0      9ɐ.00 ø   ĕȥʏǥƃ ūͮΫ 7ü˳.00
2Ͷχʄ020-04-14 ï     ǟ1.00 ʶ    ̀  Ŭƛ2.0\x9fν0Ġ      1.00     ͛ 6.00ɇ
ɟĖ20ɮ2ͺō̪ȕͿ0-ˀ0ŋ4ʥȧ͕-15    ĝ  5.0˃ɷ0ó É     7.j00     Ȩ̺ 4.00    ƀ  7ŅƂϭ.0(¬˩"0
2020-04-16  ƿ şǏ   8.0Ɓ0      6.00ɔ Ɛ     2Ť.00 Ĵ  Πǥ ³  0Ͽ.0Ɓǡ0"""

    def __init__(s_elf, iterations: Optional[int]=None, depth: Optional[int]=None, learning_rat: Optional[float]=None, logg: Optional[st]='Silent', l2_leaf: Optional[float]=None, thread_co_unt: Optional[int]=None, **kwargs):
        s_elf.iterations = iterations
        s_elf.depth = depth
        s_elf.learning_rate = learning_rat
        s_elf.logging_level = logg
        s_elf.l2_leaf_reg = l2_leaf
        s_elf.thread_count = thread_co_unt
        s_elf.kwargs = kwargs
        supe().__init__(iterations=iterations, depth=depth, learning_rate=learning_rat, logging_level=logg, thread_count=thread_co_unt, l2_leaf_reg=l2_leaf, **kwargs)
